# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"Mem0ai to API","version":"0.1.0"},"paths":{"/authorized":{"get":{"tags":["mem0ai"],"summary":"Authorized","description":"Check authorization status","operationId":"authorized_authorized_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/store":{"post":{"tags":["mem0ai"],"summary":"Store Memory","description":"Create a new memory.","operationId":"store_memory_store_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StoreMemoryData"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessfulResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/update/{memory_id}":{"put":{"tags":["mem0ai"],"summary":"Update Memory","description":"Update a memory by ID.","operationId":"update_memory_update__memory_id__put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"memory_id","in":"path","required":true,"schema":{"type":"string","title":"Memory Id"}},{"name":"data","in":"query","required":true,"schema":{"type":"string","title":"Data"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/search":{"post":{"tags":["mem0ai"],"summary":"Search Memories","description":"Search for memories.","operationId":"search_memories_search_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SearchMemoryData"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/retrieve":{"get":{"tags":["mem0ai"],"summary":"Retrieve Memories","description":"List all memories.","operationId":"retrieve_memories_retrieve_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"user_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"User Id"}},{"name":"agent_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Agent Id"}},{"name":"run_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Run Id"}},{"name":"limit","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/retrieve/{memory_id}":{"get":{"tags":["mem0ai"],"summary":"Retrieve Memory","description":"Get the history of changes for a memory by ID.","operationId":"retrieve_memory_retrieve__memory_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"memory_id","in":"path","required":true,"schema":{"type":"string","title":"Memory Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/delete/{memory_id}":{"delete":{"tags":["mem0ai"],"summary":"Delete Memory","description":"Delete a memory by ID.","operationId":"delete_memory_delete__memory_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"memory_id","in":"path","required":true,"schema":{"type":"string","title":"Memory Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/delete-all":{"delete":{"tags":["mem0ai"],"summary":"Delete Memory","description":"Delete a memory by ID.","operationId":"delete_memory_delete_all_delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"user_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"User Id"}},{"name":"agent_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Agent Id"}},{"name":"run_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Run Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/reset-all":{"delete":{"tags":["mem0ai"],"summary":"Reset All Memories","description":"Reset the memory store.","operationId":"reset_all_memories_reset_all_delete","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"SearchMemoryData":{"properties":{"query":{"type":"string","title":"Query"},"user_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"User Id"},"agent_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Agent Id"},"run_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Run Id"},"limit":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Limit","default":100},"filters":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Filters"}},"type":"object","required":["query"],"title":"SearchMemoryData"},"StoreMemoryData":{"properties":{"data":{"type":"string","title":"Data"},"user_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"User Id"},"agent_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Agent Id"},"run_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Run Id"},"metadata":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Metadata"},"filters":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Filters"},"prompt":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Prompt"}},"type":"object","required":["data"],"title":"StoreMemoryData"},"SuccessfulResponse":{"properties":{"code":{"type":"integer","title":"Code","default":0},"data":{"title":"Data"}},"type":"object","title":"SuccessfulResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}